requirements.txt
# ehrx - EHR Extraction Tool Requirements (Post-LayoutParser Cleanup)
# 
# SYSTEM DEPENDENCIES:
# - poppler (for PDF processing via pdf2image)
# - tesseract (for OCR text extraction)

# Core processing
pytesseract>=0.3.0
opencv-python>=4.9.0
numpy<2

# PDF processing
pymupdf==1.23.*
pdf2image>=1.17.0

# Data handling
pandas>=2.2.0
pyyaml>=6.0.0
pydantic>=2.6.0

# VLM integration (Google Cloud Vertex AI + Gemini)
google-cloud-aiplatform>=1.38.0
vertexai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0

# Web server (for Cloud Run deployment)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Development
pytest>=8.0.0
pytest-cov>=4.1.0

